+++ app/Http/Controllers/AcademicDevelopment/StudentAwardController.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\Controller;
 use App\Models\StudentAwardDocument;
 use Illuminate\Support\Facades\Storage;
 use App\Models\FormBuilder\AcademicDevelopmentForm;
+use App\Http\Controllers\Maintenances\LockController;
 
 class StudentAwardController extends Controller
 {
@@ -22,7 +23,9 @@ class StudentAwardController extends Controller
     {
         $this->authorize('viewAny', StudentAward::class);
 
-        $student_awards = StudentAward::where('user_id', auth()->id())->orderBy('student_awards.updated_at', 'desc')->get();
+        $student_awards = StudentAward::where('user_id', auth()->id())
+                            ->select(DB::raw('student_awards.*, QUARTER(student_awards.updated_at) as quarter'))
+                            ->orderBy('student_awards.updated_at', 'desc')->get();
         return view('academic-development.student-awards.index', compact('student_awards'));
     }
 
@@ -58,6 +61,9 @@ class StudentAwardController extends Controller
 
         $student_award = StudentAward::create($input);
         $student_award->update(['user_id' => auth()->id()]);
+
+        $string = str_replace(' ', '-', $request->input('description')); // Replaces all spaces with hyphens.
+        $description =  preg_replace('/[^A-Za-z0-9\-]/', '', $string); // Removes special chars.
         
         if($request->has('document')){
             
@@ -68,7 +74,7 @@ class StudentAwardController extends Controller
                     $temporaryPath = "documents/tmp/".$document."/".$temporaryFile->filename;
                     $info = pathinfo(storage_path().'/documents/tmp/'.$document."/".$temporaryFile->filename);
                     $ext = $info['extension'];
-                    $fileName = 'StudentAward-'.$request->input('description').'-'.now()->timestamp.uniqid().'.'.$ext;
+                    $fileName = 'StudentAward-'.$description.'-'.now()->timestamp.uniqid().'.'.$ext;
                     $newPath = "documents/".$fileName;
                     Storage::move($temporaryPath, $newPath);
                     Storage::deleteDirectory("documents/tmp/".$document);
@@ -119,6 +125,11 @@ class StudentAwardController extends Controller
 
         if(AcademicDevelopmentForm::where('id', 3)->pluck('is_active')->first() == 0)
             return view('inactive');
+
+        if(LockController::isLocked($student_award->id, 18)){
+            return redirect()->back()->with('cannot_access', 'Cannot be edited.');
+        }
+
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(3)");
 
         $documents = StudentAwardDocument::where('student_award_id', $student_award->id)->get()->toArray();
@@ -143,7 +154,12 @@ class StudentAwardController extends Controller
             return view('inactive');
         $input = $request->except(['_token', '_method', 'document']);
 
+        $student_award->update(['description' => '-clear']);
+
         $student_award->update($input);
+
+        $string = str_replace(' ', '-', $request->input('description')); // Replaces all spaces with hyphens.
+        $description =  preg_replace('/[^A-Za-z0-9\-]/', '', $string); // Removes special chars.
         
         if($request->has('document')){
             
@@ -154,7 +170,7 @@ class StudentAwardController extends Controller
                     $temporaryPath = "documents/tmp/".$document."/".$temporaryFile->filename;
                     $info = pathinfo(storage_path().'/documents/tmp/'.$document."/".$temporaryFile->filename);
                     $ext = $info['extension'];
-                    $fileName = 'StudentAward-'.$request->input('description').'-'.now()->timestamp.uniqid().'.'.$ext;
+                    $fileName = 'StudentAward-'.$description.'-'.now()->timestamp.uniqid().'.'.$ext;
                     $newPath = "documents/".$fileName;
                     Storage::move($temporaryPath, $newPath);
                     Storage::deleteDirectory("documents/tmp/".$document);
@@ -180,6 +196,10 @@ class StudentAwardController extends Controller
     public function destroy(StudentAward $student_award)
     {
         $this->authorize('delete', StudentAward::class);
+
+        if(LockController::isLocked($student_award->id, 18)){
+            return redirect()->back()->with('cannot_access', 'Cannot be edited.');
+        }
 
         StudentAwardDocument::where('student_award_id', $student_award->id)->delete();
         $student_award->delete();
